Send local_forest_runtime stdout and stderr to DEVNULL
When sending quilc and qvm stdout and stderr to a subprocess.PIPE,
it's possible for the pipe buffer to fill up and cause the processes
to hang when attempting to log on stderr. Switch these to
subprocess.DEVNULL instead. This means you can no longer read log
messages from the handles, but has the advantage that it doesn't
require user intervention to periodically drain the pipe buffer to
prevent a hang.

Closes #1110